% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{decimalplaces}
\alias{decimalplaces}
\title{Find the decimal places in a float}
\usage{
decimalplaces(x)
}
\arguments{
\item{x}{A float.}
}
\value{
The number of decimal places required to represent the float.
}
\description{
Find the decimal places in a float
}
