% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailytemps.R
\name{dt_modifiedlp}
\alias{dt_modifiedlp}
\title{Generate a daily temperature matrix based on a modified version of the Logan-Parton model.}
\usage{
dt_modifiedlp(
  dtr,
  tbar,
  resolution = 1,
  tempdecimalprecision = 1,
  clamp = TRUE
)
}
\arguments{
\item{dtr}{Daily temperature range.}

\item{tbar}{A vector of mean temperatures of the day.}

\item{resolution}{Temporal resolution of the output (in terms of hours).}

\item{tempdecimalprecision}{Decimal precision of the resultant temperature.}

\item{clamp}{Whether to clamp the output values between the minimum and maximum values of \code{tbar}.}
}
\value{
A matrix of \code{24/resolution} (rows) x \code{length(tbar)} (cols) containing temperature values at given times.
}
\description{
Generate a daily temperature matrix based on a modified version of the Logan-Parton model.
}
\note{
If \code{length(tbar) == 1}, clamp is automatically set to \code{FALSE}.
}
\section{Modifications}{

This is a heavily restricted and modified version of the Logan-Parton model. It is parameterised to a day length of 12 and a sunrise time of 6am.

It makes sure that the mean and median of a given column equals the associated \code{tbar}
}

\examples{
dt_modifiedlp(10, tbar = seq(20,45,0.1))
}
\references{
Parton, W.J. and Logan, J.A., 1981. A model for diurnal variation in soil and air temperature. Agricultural meteorology, 23, pp.205-216.
}
