% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailytemps.R
\name{dt_tri}
\alias{dt_tri}
\title{Generate a simple daily temperature triangle curve}
\usage{
dt_tri(
  dtr,
  tbar,
  sunrise = 6,
  resolution = 1,
  tempdecimalprecision = 1,
  clamp = TRUE
)
}
\arguments{
\item{dtr}{Daily temperature range.}

\item{tbar}{A vector of target mean temperatures.}

\item{sunrise}{The time that the sun rises.}

\item{resolution}{The resolution of the time series in hours.}

\item{tempdecimalprecision}{Decimal precision of the resultant temperature.}

\item{clamp}{Whether to clamp the output between \code{min(tbar)} and \code{max(tbar)}, defaults to \code{TRUE}.}
}
\value{
A matrix of \code{24/resolution} (rows) x \code{length(tbar)} (cols) containing temperature values at given times.
}
\description{
Generate a simple daily temperature triangle curve
}
\note{
If \code{length(tbar) == 1}, clamp is automatically set to \code{FALSE}.
}
\examples{
dt_tri(dtr = 15, tbar = seq(20,45,0.1), sunrise = 6, resolution = 0.5, clamp = TRUE)
}
