% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_dps}
\alias{find_dps}
\title{Find the decimal places in a vector of floats}
\usage{
find_dps(v)
}
\arguments{
\item{v}{A vector of floats.}
}
\value{
A vector of the number of decimal places required to represent each float.
}
\description{
Find the decimal places in a vector of floats
}
