% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nla-package.R
\docType{package}
\name{nla-package}
\alias{nla}
\alias{nla-package}
\title{nla: Non-Linear Averaging Tools}
\description{
Perform non-linear averaging over time in a fast and repeatable manner.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fwimp/nla}
  \item \url{https://fwimp.github.io/nla/}
  \item Report bugs at \url{https://github.com/fwimp/nla/issues}
}

}
\author{
\strong{Maintainer}: Francis Windram \email{francis.windram17@imperial.ac.uk} (\href{https://orcid.org/0000-0002-2129-826X}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Marta Shocket \email{m.shocket@lancaster.ac.uk} (\href{https://orcid.org/0000-0002-8995-4446}{ORCID})
}

}
\keyword{internal}
