% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nla.R
\name{nla_tpc_analytic}
\alias{nla_tpc_analytic}
\title{Run rate summation on a TPC with a functional form}
\usage{
nla_tpc_analytic(
  tpc_func,
  paramlist,
  dailytemps,
  tempaxis_name = "temp",
  minvalue = NULL,
  maxvalue = NULL,
  default_value = NULL
)
}
\arguments{
\item{tpc_func}{The tpc function to apply.}

\item{paramlist}{A named list of parameters (or vectors of parameters) for \code{tpc_func}. If \code{paramlist$labels} is provided, this will be used as the rownames of the resultant matrix.}

\item{dailytemps}{A matrix of temperature across a 24 hour period at various mean temperatures.}

\item{tempaxis_name}{Override the name of the temperature axis within the tpc func (if it is not "temp", the default in rTPC).}

\item{minvalue}{Minimum value a TPC can have (acts as a lower clamp).}

\item{maxvalue}{Maximum value a TPC can have (acts as a lower clamp).}

\item{default_value}{Value to replace NaN results from the TPC with, should these occur (this may be changed later to a possible function instead).}
}
\value{
A matrix of averaged values, where columns are temperatures, and rows are individual tpc runs.
}
\description{
Run rate summation on a TPC with a functional form
}
\examples{
test_param_list <- list(
"a" = c(0.01, 0.02),
"b" = 2,
"tmin" = 10,
"tmax" = 30,
"labels" = c("test", "test2")
)
test_dt <- dt_sin(dtr = 15, tbar = seq(20,45,0.1), sunrise = 6, resolution = 0.5, clamp = TRUE)
nla_tpc_analytic(tpc_func = rTPC::briere2_1999, paramlist = test_param_list, dailytemps = test_dt, minvalue = 0, default_value = 0)

}
