% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nla.R
\name{nla_tpc_empirical}
\alias{nla_tpc_empirical}
\title{Perform nonlinear averaging of a temperature-dependent variable across a range of temperatures from empirically-derived data}
\usage{
nla_tpc_empirical(tpcs, dailytemps, tempaxis)
}
\arguments{
\item{tpcs}{A matrix of values measured across a temperature axis (represented in \code{tempaxis}).}

\item{dailytemps}{A matrix of temperatures across a 24h period, equating to each entry in \code{tempaxis}.}

\item{tempaxis}{The basis temperature axis.}
}
\value{
A matrix of averaged values, where columns are temperatures, and rows are individual tpc runs.
}
\description{
Perform nonlinear averaging of a temperature-dependent variable across a range of temperatures from empirically-derived data
}
\note{
The precision of \code{tempaxis} MUST be greater than or equal to the values of \code{dailytemps}.

The width of the \code{tpcs} matrix MUST be the same length as tempaxis.
}
\section{Input dimensionality}{


This function relies on carefully and appropriately-formatted inputs.

\code{tpcs} should be a 2D matrix where the rows correspond to individual tpcs or bootstrap iterations of a tpc, and columns correspond to the temperature axis in \code{tempaxis}.

\code{dailytemps} should be a 2D matrix where the rows correspond to sampling times across a 24h window (so if your temperatures are hourly, this should have 24 rows), and columns correspond to mean temperatures. So if column 5 represents 10°C, \code{mean(dailytemps[,5]) == 10}
}

\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
nla_tpc_empirical()
\dontshow{\}) # examplesIf}
}
